CRUD

+++ TechStoreEll.Api/Models/User.cs
@@ -21,7 +21,7 @@ public partial class User
 
     public string? Phone { get; set; }
 
-    public short RoleId { get; set; }
+    public int RoleId { get; set; }
 
     public bool IsActive { get; set; }
